requirements.txt
fastapi 
uvicorn 
pymongo 
groq 
cartesia 
python-dotenv
python-multipart
ffmpeg